+++ src/daft-local-execution/src/intermediate_ops/intermediate_op.rs
@@ -13,9 +13,7 @@ use crate::{
     channel::{
         create_channel, create_single_channel, MultiReceiver, MultiSender, SingleReceiver,
         SingleSender,
-    },
-    pipeline::PipelineNode,
-    ExecutionRuntimeHandle, NUM_CPUS,
+    }, pipeline::PipelineNode, runtime_stats::RuntimeStatsContext, ExecutionRuntimeHandle, NUM_CPUS
 };
 
 use super::state::OperatorTaskState;
@@ -24,75 +22,7 @@ pub trait IntermediateOperator: Send + Sync {
     fn name(&self) -> &'static str;
 }
 
-#[derive(Default)]
-struct RuntimeStatsContext {
-    name: String,
-    rows_received: AtomicU64,
-    rows_emitted: AtomicU64,
-    cpu_us: AtomicU64,
-}
 
-#[derive(Debug)]
-struct RuntimeStats {
-    rows_received: u64,
-    rows_emitted: u64,
-    cpu_us: u64,
-}
-
-impl RuntimeStatsContext {
-    fn new(name: String) -> Self {
-        Self {
-            name,
-            rows_received: AtomicU64::new(0),
-            rows_emitted: AtomicU64::new(0),
-            cpu_us: AtomicU64::new(0),
-        }
-    }
-    fn in_span<F: FnOnce() -> T, T>(&self, span: &tracing::Span, f: F) -> T {
-        let _enter = span.enter();
-        let start = Instant::now();
-        let result = f();
-        let total = start.elapsed();
-        let micros = total.as_micros() as u64;
-        self.cpu_us
-            .fetch_add(micros, std::sync::atomic::Ordering::Relaxed);
-        result
-    }
-
-    fn mark_rows_received(&self, rows: u64) {
-        self.rows_received
-            .fetch_add(rows, std::sync::atomic::Ordering::Relaxed);
-    }
-
-    fn mark_rows_emitted(&self, rows: u64) {
-        self.rows_emitted
-            .fetch_add(rows, std::sync::atomic::Ordering::Relaxed);
-    }
-
-    fn reset(&self) {
-        self.rows_received
-            .store(0, std::sync::atomic::Ordering::Release);
-        self.rows_emitted
-            .store(0, std::sync::atomic::Ordering::Release);
-        self.cpu_us.store(0, std::sync::atomic::Ordering::Release);
-    }
-
-    fn result(&self) -> RuntimeStats {
-        RuntimeStats {
-            rows_received: self
-                .rows_received
-                .load(std::sync::atomic::Ordering::Relaxed),
-            rows_emitted: self.rows_emitted.load(std::sync::atomic::Ordering::Relaxed),
-            cpu_us: self.cpu_us.load(std::sync::atomic::Ordering::Relaxed),
-        }
-    }
-}
-
-impl Drop for RuntimeStatsContext {
-    fn drop(&mut self) {
-        println!("{}={:?}", self.name, self.result());
-    }
-}
 
 pub(crate) struct IntermediateNode {
     intermediate_op: Arc<dyn IntermediateOperator>,
@@ -132,16 +62,18 @@ impl IntermediateNode {
             state.add(result);
             if let Some(part) = state.try_clear() {
                 let part = part?;
-                rt_context.mark_rows_emitted(part.len() as u64);
-
+                let len = part.len();
                 let _ = sender.send(part).await;
+                rt_context.mark_rows_emitted(len as u64);
+
             }
         }
         if let Some(part) = state.clear() {
             let part = part?;
-            rt_context.mark_rows_emitted(part.len() as u64);
-
+            let len = part.len();
             let _ = sender.send(part).await;
+            rt_context.mark_rows_emitted(len as u64);
+
         }
         Ok(())
     }
