+++ src/daft-local-execution/src/sinks/blocking_sink.rs
@@ -5,9 +5,7 @@ use daft_micropartition::MicroPartition;
 use tracing::info_span;
 
 use crate::{
-    channel::{create_channel, MultiSender},
-    pipeline::PipelineNode,
-    ExecutionRuntimeHandle, NUM_CPUS,
+    channel::{create_channel, MultiSender}, pipeline::PipelineNode, runtime_stats::RuntimeStatsContext, ExecutionRuntimeHandle, NUM_CPUS
 };
 use async_trait::async_trait;
 pub enum BlockingSinkStatus {
@@ -19,7 +17,6 @@ pub enum BlockingSinkStatus {
 pub trait BlockingSink: Send + Sync {
     fn sink(&mut self, input: &Arc<MicroPartition>) -> DaftResult<BlockingSinkStatus>;
     fn finalize(&mut self) -> DaftResult<Option<Arc<MicroPartition>>>;
-    #[allow(dead_code)]
     fn name(&self) -> &'static str;
 }
 
@@ -28,6 +25,7 @@ pub(crate) struct BlockingSinkNode {
     op: Arc<tokio::sync::Mutex<Box<dyn BlockingSink>>>,
     name: &'static str,
     child: Box<dyn PipelineNode>,
+    runtime_stats: Arc<RuntimeStatsContext>,
 }
 
 impl BlockingSinkNode {
@@ -37,6 +35,7 @@ impl BlockingSinkNode {
             op: Arc::new(tokio::sync::Mutex::new(op)),
             name,
             child,
+            runtime_stats: Arc::new(RuntimeStatsContext::new(name.to_string()))
         }
     }
     pub(crate) fn boxed(self) -> Box<dyn PipelineNode> {
@@ -64,18 +63,23 @@ impl PipelineNode for BlockingSinkNode {
         let child = self.child.as_mut();
         child.start(sender, runtime_handle).await?;
         let op = self.op.clone();
+
+        let rt_context = self.runtime_stats.clone();
         runtime_handle.spawn(async move {
             let span = info_span!("BlockingSinkNode::execute");
             let mut guard = op.lock().await;
             while let Some(val) = streaming_receiver.recv().await {
-                if let BlockingSinkStatus::Finished = span.in_scope(|| guard.sink(&val))? {
+                rt_context.mark_rows_received(val.len() as u64);
+                if let BlockingSinkStatus::Finished = rt_context.in_span(&span, || guard.sink(&val))? {
                     break;
                 }
             }
             let finalized_result =
-                info_span!("BlockingSinkNode::finalize").in_scope(|| guard.finalize())?;
+                rt_context.in_span(&info_span!("BlockingSinkNode::finalize"), || guard.finalize())?;
             if let Some(part) = finalized_result {
+                let len = part.len();
                 let _ = destination.get_next_sender().send(part).await;
+                rt_context.mark_rows_emitted(len as u64);
             }
             Ok(())
         });
